+++ ui/app/reducers/app.js
@@ -317,6 +317,16 @@ function reduceApp (state, action) {
         isLoading: false,
       })
 
+    case actions.SHOW_SUB_LOADING_INDICATION:
+      return extend(appState, {
+        isSubLoading: true,
+      })
+
+    case actions.HIDE_SUB_LOADING_INDICATION:
+      return extend(appState, {
+        isSubLoading: false,
+      })
+
     case actions.CLEAR_SEED_WORD_CACHE:
       return extend(appState, {
         transForward: true,
@@ -369,7 +379,7 @@ function reduceApp (state, action) {
         },
       })
 
-    case actions.SHOW_ETH_WARNING:
+    case actions.BUY_ETH_SUBVIEW:
       return extend(appState, {
         transForward: true,
         currentView: {
@@ -377,9 +387,103 @@ function reduceApp (state, action) {
           context: appState.currentView.context,
         },
         accountDetail: {
-          subview: 'buy-eth-warning',
+          subview: 'buyForm',
+          amount: '5.00',
+          buyAddress: appState.currentView.context,
+          formView: {
+            coinbase: true,
+            shapeshift: false,
+          },
         },
       })
+
+    case actions.UPDATE_BUY_ADDRESS:
+      return extend(appState, {
+        accountDetail: {
+          subview: 'buyForm',
+          formView: {
+            coinbase: true,
+            shapeshift: false,
+          },
+          buyAddress: action.value,
+          amount: appState.accountDetail.amount,
+        },
+      })
+
+    case actions.UPDATE_COINBASE_AMOUNT:
+      return extend(appState, {
+        accountDetail: {
+          subview: 'buyForm',
+          formView: {
+            coinbase: true,
+            shapeshift: false,
+          },
+          buyAddress: appState.accountDetail.buyAddress,
+          amount: action.value,
+        },
+      })
+
+    case actions.COINBASE_SUBVIEW:
+      return extend(appState, {
+        accountDetail: {
+          subview: 'buyForm',
+          formView: {
+            coinbase: true,
+            shapeshift: false,
+          },
+          buyAddress: appState.accountDetail.buyAddress,
+          amount: appState.accountDetail.amount,
+        },
+      })
+
+    case actions.SHAPESHIFT_SUBVIEW:
+      return extend(appState, {
+        accountDetail: {
+          subview: 'buyForm',
+          formView: {
+            coinbase: false,
+            shapeshift: true,
+            marketinfo: action.value.marketinfo,
+            coinOptions: action.value.coinOptions,
+          },
+          buyAddress: appState.accountDetail.buyAddress,
+          amount: appState.accountDetail.amount,
+        },
+      })
+
+    case actions.PAIR_UPDATE:
+      return extend(appState, {
+        accountDetail: {
+          subview: 'buyForm',
+          formView: {
+            coinbase: false,
+            shapeshift: true,
+            marketinfo: action.value.marketinfo,
+            coinOptions: appState.accountDetail.formView.coinOptions,
+          },
+          buyAddress: appState.accountDetail.buyAddress,
+          amount: appState.accountDetail.amount,
+          warning: null,
+        },
+      })
+
+    case actions.COIN_SHIFT_REQUEST:
+      return extend(appState, {
+        accountDetail: {
+          subview: 'buyForm',
+          formView: {
+            coinbase: false,
+            shapeshift: true,
+            marketinfo: appState.accountDetail.formView.marketinfo,
+            coinOptions: appState.accountDetail.formView.coinOptions,
+            response: action.value.response,
+          },
+          buyAddress: appState.accountDetail.buyAddress,
+          amount: appState.accountDetail.amount,
+          warning: null,
+        },
+      })
+
     default:
       return appState
   }
@@ -404,5 +508,4 @@ function indexForPending (state, txId) {
   })
   return idx
 }
-