+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Current Master
 
+- Add ShapeShift integration
+- Add a for for Coinbase Buy button to support different amounts other then $5
 - Fix various typos.
 
 ## 2.7.3 2016-07-29